Removed ColorSheet interfaces
