--- src/main/java/org/reactome/server/tools/diagram/exporter/raster/profiles/ColorSheet.java
@@ -1,9 +1,0 @@
-package org.reactome.server.tools.diagram.exporter.raster.profiles;
-
-/**
- * Represents a Reactome color sheet or color profile.
- */
-public interface ColorSheet {
-
-	String getName();
-}
