--- src/main/java/org/reactome/server/tools/diagram/exporter/raster/profiles/EnrichmentSheetImpl.java
@@ -1,25 +1,0 @@
-package org.reactome.server.tools.diagram.exporter.raster.profiles;
-
-
-import java.awt.*;
-
-public class EnrichmentSheetImpl implements EnrichmentSheet {
-
-	private GradientSheetImpl gradient;
-	private Color text;
-
-	@Override
-	public Color getText() {
-		return text;
-	}
-
-	@Override
-	public GradientSheet getGradient() {
-		return gradient;
-	}
-
-	public void setText(String color) {
-		this.text = ColorFactory.parseColor(color);
-	}
-
-}
