--- src/main/java/org/reactome/server/tools/diagram/exporter/raster/profiles/ExpressionSheetImpl.java
@@ -1,10 +1,0 @@
-package org.reactome.server.tools.diagram.exporter.raster.profiles;
-
-public class ExpressionSheetImpl extends EnrichmentSheetImpl implements ExpressionSheet {
-	private LegendSheetImpl legend;
-
-	@Override
-	public LegendSheet getLegend() {
-		return legend;
-	}
-}
