+++ src/main/java/org/reactome/server/tools/diagram/exporter/raster/profiles/GradientSheet.java
@@ -1,13 +1,38 @@
 package org.reactome.server.tools.diagram.exporter.raster.profiles;
 
+
 import java.awt.*;
 
-public interface GradientSheet {
+public class GradientSheet {
+	private Color min;
+	private Color stop;
+	private Color max;
 
-	Color getMin();
 
-	Color getStop();
+	public Color getMin() {
+		return min;
+	}
 
-	Color getMax();
+	public void setMin(String color) {
+		this.min = ColorFactory.parseColor(color);
+	}
+
+
+	public Color getStop() {
+		return stop;
+	}
+
+	public void setStop(String color) {
+		this.stop = ColorFactory.parseColor(color);
+	}
+
+
+	public Color getMax() {
+		return max;
+	}
+
+	public void setMax(String color) {
+		this.max = ColorFactory.parseColor(color);
+	}
 
 }
