+++ src/main/java/org/reactome/server/tools/diagram/exporter/raster/profiles/PropertiesColorSheet.java
@@ -2,26 +2,96 @@ package org.reactome.server.tools.diagram.exporter.raster.profiles;
 
 import java.awt.*;
 
-/**
- *
- */
-public interface PropertiesColorSheet {
+public class PropertiesColorSheet {
+	private Color halo;
+	private Color flag;
+	private Color disease;
+	private Color selection;
+	private Color hovering;
+	private Color highlight;
+	private Color text;
+	private Color button;
+	private Color trigger;
 
-	Color getHalo();
+	
+	public Color getHalo() {
+		return halo;
+	}
 
-	Color getFlag();
+	
+	public Color getFlag() {
+		return flag;
+	}
 
-	Color getDisease();
+	
+	public Color getDisease() {
+		return disease;
+	}
 
-	Color getSelection();
+	
+	public Color getSelection() {
+		return selection;
+	}
 
-	Color getHovering();
+	
+	public Color getHovering() {
+		return hovering;
+	}
 
-	Color getHighlight();
+	
+	public Color getHighlight() {
+		return highlight;
+	}
 
-	Color getText();
+	
+	public Color getText() {
+		return text;
+	}
 
-	Color getButton();
+	
+	public Color getButton() {
+		return button;
+	}
 
-	Color getTrigger();
+	
+	public Color getTrigger() {
+		return trigger;
+	}
+
+	public void setHalo(String color) {
+		this.halo = ColorFactory.parseColor(color);
+	}
+
+	public void setFlag(String color) {
+		this.flag = ColorFactory.parseColor(color);
+	}
+
+	public void setDisease(String color) {
+		this.disease = ColorFactory.parseColor(color);
+	}
+
+	public void setSelection(String color) {
+		this.selection = ColorFactory.parseColor(color);
+	}
+
+	public void setHovering(String color) {
+		this.hovering = ColorFactory.parseColor(color);
+	}
+
+	public void setHighlight(String color) {
+		this.highlight = ColorFactory.parseColor(color);
+	}
+
+	public void setText(String color) {
+		this.text = ColorFactory.parseColor(color);
+	}
+
+	public void setButton(String color) {
+		this.button = ColorFactory.parseColor(color);
+	}
+
+	public void setTrigger(String color) {
+		this.trigger = ColorFactory.parseColor(color);
+	}
+
 }
