+++ src/main/java/org/reactome/server/tools/diagram/exporter/raster/profiles/ThumbnailColorSheet.java
@@ -2,14 +2,56 @@ package org.reactome.server.tools.diagram.exporter.raster.profiles;
 
 import java.awt.*;
 
-public interface ThumbnailColorSheet {
-	Color getNode();
+public class ThumbnailColorSheet {
+	private Color node;
+	private Color edge;
+	private Color hovering;
+	private Color highlight;
+	private Color selection;
 
-	Color getEdge();
+	
+	public Color getNode() {
+		return node;
+	}
 
-	Color getHovering();
+	public void setNode(String color) {
+		this.node = ColorFactory.parseColor(color);
+	}
 
-	Color getHighlight();
+	
+	public Color getEdge() {
+		return edge;
+	}
 
-	Color getSelection();
+	public void setEdge(String color) {
+		this.edge = ColorFactory.parseColor(color);
+	}
+
+	
+	public Color getHovering() {
+		return hovering;
+	}
+
+	public void setHovering(String color) {
+		this.hovering = ColorFactory.parseColor(color);
+	}
+
+	
+	public Color getHighlight() {
+		return highlight;
+	}
+
+	public void setHighlight(String color) {
+		this.highlight = ColorFactory.parseColor(color);
+	}
+
+	
+	public Color getSelection() {
+		return selection;
+	}
+
+	public void setSelection(String color) {
+		this.selection = ColorFactory.parseColor(color);
+	}
+
 }
